backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0

# Authentication JWT
djangorestframework-simplejwt==5.3.1

# PostgreSQL Database
psycopg2-binary==2.9.9

# Environment Variables
django-environ==0.11.2

# CORS Headers (para desarrollo con frontend separado)
django-cors-headers==4.3.1

# Additional utilities
python-decouple==3.8

# Image processing (ImageField)
Pillow==10.4.0
